Fix CI issues: move context to separate file and fix imports
