+++ frontend/src/components/slack/index.ts
@@ -4,4 +4,5 @@ export { default as WorkspaceList } from './WorkspaceList'
 export { default as ChannelList } from './ChannelList'
 export { default as MessageList } from './MessageList'
 export { default as ThreadView } from './ThreadView'
-export { default as SlackUserDisplay, SlackUserCacheProvider } from './SlackUserDisplay'
+export { default as SlackUserDisplay } from './SlackUserDisplay'
+export { SlackUserCacheProvider, SlackUser } from './SlackUserContext'
